+++ onnx/graph.go
@@ -343,10 +343,8 @@ func (m *Model) convertNode(ctx *context.Context, g *Graph, node *protos.NodePro
 	case "DynamicQuantizeLinear":
 		res = convertDynamicQuantizeLinear(convertedOutputs, node, inputs)
 	case "Trilu":
-		fmt.Println("entered trilu")
 		res = convertTrilu(m, convertedOutputs, node, inputs)
 	case "ScatterND":
-		fmt.Println("entered ScatterND")
 		res = convertScatterND(m, convertedOutputs, node, inputs)
 
 		// Ops not implemented: